// modulo5/arquitetura-de-software-2/atividade-arquitetura-II/src/model/user.ts
export enum USER_ROLES {
  NORMAL = 'NORMAL',
  ADMIN = 'ADMIN',
}

export type authenticationData = {
  id: string;
  role: USER_ROLES;
};

export type userData = {
  name: string;
  nickname: string;
  email: string;
  password: string;
  role: USER_ROLES;
};

export interface signupInputDTO {
  name: string;
  nickname: string;
  email: string;
  password: string;
  role: USER_ROLES;
}

export type user = userData & { id: string };